Adds db_models module with ORM table definitions for transactions and updates, and introduces test fixtures for database setup in test_utils. Moves BaseTable db table definition classes to db_models. Refactors tests to reflect new module structure. Adds conftest.py test fixture config file.
