+++ .gitignore
@@ -179,3 +179,6 @@ cython_debug/
 
 # PyPI configuration file
 .pypirc
+
+# Claude Code markdown file
+CLAUDE.md
