+++ tests/test_utils/test_db_utils.py
@@ -2,74 +2,37 @@
 """
 budgy.utils.tests.test_utils.test_db_utils
 
-Tests for budgy.utils.db_utils module.
+Tests for budgy.utils.db_utils module — CSV processing and database upload functions.
 """
 # Standard library imports
 from datetime import datetime
-from collections import defaultdict
 import os
-import pytest
-import sys
-
-# Third-party imports
-import pandas as pd
 
 # Local imports
+from local_db import DuplicateError
+from budgy.utils.db_models import TransactionsTable, TableStatus
 from budgy.utils.db_utils import (
-    DatabaseManager, DatabaseFile, DuplicateError,
-    TableStatus, transactions_table_manager, update_table_manager, 
-    TransactionsTable, UpdatesTable, generate_update_entry,
-    generate_base_hash,
+    generate_update_entry, generate_base_hash,
     iter_csv_not_uploaded, iter_val_csv_file,
     update_categories_if_diff, upload_csv_to_db, columns
 )
+from tests.test_utils.conftest import TEST_CSV_DIR
 
 # Initialize module logger
 import logging
 logger = logging.getLogger(__name__)
 
-TEST_CSV_DIR = os.path.join(os.getcwd(), "tests", "test_utils", "test_csv_download_files")
-TEST_DB_DIR = os.path.join(os.getcwd(), "tests", "test_utils", "test_database")
-TEST_DB_FILENAME = "test_database.db"
-TEST_DB_FILEPATH = os.path.join(TEST_DB_DIR, TEST_DB_FILENAME)
-
-test_db_file = DatabaseFile(TEST_DB_FILEPATH, TEST_DB_DIR)
-test_transaction_manager = DatabaseManager(TransactionsTable, test_db_file)
-test_updates_manager = DatabaseManager(UpdatesTable, test_db_file)
-
-update_items = [
-    {
-        "timestamp": datetime(2024, 1, 1),
-        "filepath": os.path.join(TEST_CSV_DIR, "TEST_UPDATE.csv"),
-        "status": TableStatus.COMPLETE
-    },
-    {
-        "timestamp": datetime(2024, 1, 2),
-        "filepath": os.path.join(TEST_CSV_DIR, "transactions_1.csv"),
-        "status": TableStatus.COMPLETE
-    },
-    {
-        "timestamp": datetime(2024, 1, 3),
-        "filepath": os.path.join(TEST_CSV_DIR, "transactions_2.csv"),
-        "status": TableStatus.INCOMPLETE
-    },
-    {
-        "timestamp": datetime(2024, 1, 5),
-        "filepath": os.path.join(TEST_CSV_DIR, "hsbifunsdovns.csv"),
-        "status": "Error - hdchiboenc"
-    }
-]
 
 duplicate_update_item = {
-        "timestamp": datetime(2024, 1, 1),
-        "filepath": os.path.join(TEST_CSV_DIR, "TEST_UPDATE.csv"),
-        "status": TableStatus.COMPLETE
-    }
+    "timestamp": datetime(2024, 1, 1),
+    "filepath": os.path.join(TEST_CSV_DIR, "TEST_UPDATE.csv"),
+    "status": TableStatus.COMPLETE
+}
 
 new_update_item = {
-        "timestamp": datetime(2024, 1, 1),
-        "filepath": os.path.join(TEST_CSV_DIR, "TEST_UPDATE_2.csv"),
-    }
+    "timestamp": datetime(2024, 1, 1),
+    "filepath": os.path.join(TEST_CSV_DIR, "TEST_UPDATE_2.csv"),
+}
 
 record_1 = {
     "authorized_date": datetime(2024, 1, 1),
@@ -115,94 +78,7 @@ duplicate_record_2 = {
 }
 
 
-@pytest.fixture()
-def clean_updates_database():
-    """Fixture to clean the database before and after each test"""
-    db_manager = test_updates_manager
-
-    try:
-        # Setup: Clean the database
-        db_manager.add_multiple_items(update_items)
-        yield db_manager  # Provide the db_manager to the test
-
-    except Exception:
-        db_manager.session.rollback()  # Rollback the session if an exception occurs
-        raise
-
-    finally:
-        db_manager.clear_table()
-
-        # Teardown: Ensure the session is closed
-        db_manager.end_session()
-
-@pytest.fixture()
-def clean_transactions_database():
-    """Fixture to clean the database before and after each test"""
-    db_manager = test_transaction_manager
-
-    try:
-        yield db_manager  # Provide the db_manager to the test
-
-    except Exception:
-        db_manager.session.rollback()  # Rollback the session if an exception occurs
-        raise
-
-    finally:
-        db_manager.clear_table()
-
-        # Teardown: Ensure the session is closed
-        db_manager.end_session()
-
-
-def test_db_file_creation():
-    """Test that the database files are created successfully."""
-    logger.debug("Starting test...")
-    db_file = transactions_table_manager.file
-    db_file.create()
-    assert os.path.exists(db_file.file_path), "Database file does not exist."
-
-
-def test_transactions_table_creation():
-    """Test that the Transactions table is created successfully and data could be retrieved from it."""
-    logger.debug("Starting test...")
-    transactions_table_manager.add_item(**record_1)
-    items = transactions_table_manager.fetch_all_items()
-    logger.debug(f"Fetched items from transaction_updates table:\n{items}")
-    assert items is not None, "Failed to fetch items from Transactions table."
-    assert isinstance(items, list), "Fetched items is not a list."
-
-    as_df = transactions_table_manager.to_dataframe()
-    logger.info(f"Transactions table as dataframe:\n{as_df}")
-    assert not as_df.empty, "Dataframe conversion resulted in empty dataframe."
-    assert list(as_df.columns) == (['id'] + list(record_1.keys()) + ["base_hash", "uq_hash"]), "Dataframe columns do not match expected columns."
-
-    transactions_table_manager.delete_items_by_attribute(**{"description": "TEST TRANSACTION"})
-
-
-def test_updates_table_creation():
-    """Test that the transaction_updates table is created successfully and data could be retrieved from it."""
-    logger.debug("Starting test...")
-    
-    try:
-        update_table_manager.add_item(**duplicate_update_item)
-    except Exception as e:
-        assert isinstance(e, DuplicateError)
-
-    items = update_table_manager.fetch_all_items()
-    logger.debug(f"Fetched items from transaction_updates table:\n{items}")
-    assert items is not None, "Failed to fetch items from transaction_updates table."
-    assert items is not None, "Failed to fetch items from Transactions table."
-    assert isinstance(items, list), "Fetched items is not a dataframe."
-
-    as_df = update_table_manager.to_dataframe()
-    logger.info(f"Updates table as dataframe:\n{as_df}")
-    assert not as_df.empty, "Dataframe conversion resulted in empty dataframe."
-    assert list(as_df.columns) == ['id'] + list(duplicate_update_item.keys()), "Dataframe columns do not match expected columns."
-
-    update_table_manager.delete_items_by_attribute(**{"filepath": "TEST_UPDATE.csv"})
-
-
-# ==================NOTE: This is where the basic schema tests end and the more sophisticated csv upload test cases begin.============== #
+# ==================NOTE: CSV upload function tests.============== #
 
 
 def test_generate_update_entry(clean_updates_database):
@@ -212,7 +88,7 @@ def test_generate_update_entry(clean_updates_database):
 
     except Exception as e:
         assert isinstance(e, DuplicateError)
-    
+
     finally:
         # This should execute without an error since the status is changing
         generate_update_entry(new_update_item["filepath"], TableStatus.INCOMPLETE, updates_db_manager=clean_updates_database)
@@ -239,10 +115,11 @@ def test_iter_csv__not_uploaded(clean_updates_database):
         # If an ORM object is returned, check to make sure that the status is set to incomplete.
         if db_item:
             assert db_item[0].status != TableStatus.COMPLETE
-        
+
         # Otherwise the filepath should not appear in the retrieved db values
         else:
             assert csv not in uploaded_files
+            from tests.test_utils.conftest import update_items
             assert csv not in [item["filepath"] for item in update_items]
 
 
@@ -285,7 +162,7 @@ def test_iter_csv_hash(clean_updates_database):
 def test_update_categories_if_diff(clean_transactions_database, clean_updates_database):
     """
     Tests the update_categories_if_diff function by uploading two CSV files.
-    
+
     The first CSV has original transactions with original categories.
     The second CSV has some of the same transactions with updated categories.
     The test verifies that:
@@ -295,17 +172,17 @@ def test_update_categories_if_diff(clean_transactions_database, clean_updates_da
     """
     transactions_db = clean_transactions_database
     updates_db = clean_updates_database
-    
+
     # Upload the first CSV file with original categories
     csv_original = os.path.join(TEST_CSV_DIR, "test_categories_original.csv")
     upload_csv_to_db(csv_original, transactions_db_manager=transactions_db, updates_db_manager=updates_db)
-    
+
     # Fetch the transactions after first upload
     df_after_first = transactions_db.to_dataframe()
-    
+
     # Verify initial data is in the database
     assert df_after_first.shape[0] == 3, f"Expected 3 records after first upload, got {df_after_first.shape[0]}"
-    
+
     # Store original detailed categories for verification
     original_categories = {}
     for _, row in df_after_first.iterrows():
@@ -314,29 +191,29 @@ def test_update_categories_if_diff(clean_transactions_database, clean_updates_da
             "primary": row[TransactionsTable.primary_category.name],
             "detailed": row[TransactionsTable.detailed_category.name]
         }
-    
+
     # Verify first transaction has original category (unchanged in second upload)
     whole_foods_row = df_after_first[df_after_first[TransactionsTable.description.name] == "Whole Foods Market"].iloc[0]
     assert whole_foods_row[TransactionsTable.detailed_category.name] == "Groceries"
-    
+
     # Verify second transaction has original category (will be updated)
     shell_row = df_after_first[df_after_first[TransactionsTable.description.name] == "Shell Gas Station"].iloc[0]
     assert shell_row[TransactionsTable.detailed_category.name] == "Gas"
-    
+
     # Verify third transaction has original category (will be updated)
     target_row = df_after_first[df_after_first[TransactionsTable.description.name] == "Target Store"].iloc[0]
     assert target_row[TransactionsTable.detailed_category.name] == "General Merchandise"
-    
+
     # Upload the second CSV file with updated categories
     csv_updated = os.path.join(TEST_CSV_DIR, "test_categories_updated.csv")
     upload_csv_to_db(csv_updated, transactions_db_manager=transactions_db, updates_db_manager=updates_db)
-    
+
     # Fetch the transactions after second upload
     df_after_second = transactions_db.to_dataframe()
-    
+
     # Verify we still have 3 records (no duplicates were added)
     assert df_after_second.shape[0] == 3, f"Expected 3 records after second upload, got {df_after_second.shape[0]}"
-    
+
     # Verify UNCHANGED transaction (Whole Foods with "Groceries")
     # This should NOT be modified since categories are identical
     whole_foods_after = df_after_second[df_after_second[TransactionsTable.description.name] == "Whole Foods Market"].iloc[0]
@@ -344,7 +221,7 @@ def test_update_categories_if_diff(clean_transactions_database, clean_updates_da
         "Whole Foods category should remain unchanged"
     assert whole_foods_after[TransactionsTable.primary_category.name] == "Food", \
         "Whole Foods primary category should remain unchanged"
-    
+
     # Verify UPDATED transaction 1 (Shell Gas Station)
     # Updated from "Gas" to "Auto Fuel"
     shell_after = df_after_second[df_after_second[TransactionsTable.description.name] == "Shell Gas Station"].iloc[0]
@@ -352,7 +229,7 @@ def test_update_categories_if_diff(clean_transactions_database, clean_updates_da
         f"Shell Gas Station detailed category should be updated to 'Auto Fuel', got '{shell_after[TransactionsTable.detailed_category.name]}'"
     assert shell_after[TransactionsTable.primary_category.name] == "Transportation", \
         "Shell Gas Station primary category should remain Transportation"
-    
+
     # Verify UPDATED transaction 2 (Target Store)
     # Updated from "General Merchandise" to "General Retail"
     target_after = df_after_second[df_after_second[TransactionsTable.description.name] == "Target Store"].iloc[0]
@@ -368,7 +245,7 @@ def test_upload_csv_to_db(clean_transactions_database, clean_updates_database):
     transactions_db = clean_transactions_database
     for csv in iter_csv_not_uploaded(csv_directory=TEST_CSV_DIR, updates_db_manager=updates_db):
         upload_csv_to_db(csv, transactions_db_manager=transactions_db, updates_db_manager=updates_db)
-    
+
     transactions_table = transactions_db.to_dataframe()
 
     assert not transactions_table.empty
